# backend/src/api/routes/mcp.py
"""MCP (Model Context Protocol) Tool Routes.

7 MCP Tools:
1. get_capabilities — Returns jurisdiction + doc type manifest
2. search_legal — Hybrid search (pgvector + full-text)
3. research_task — Complex research with question decomposition
4. get_document — Retrieve full document with chunks and citations
5. get_citations — Citation graph retrieval
6. check_updates — Legislative changes since date
7. jurisdiction_summary — Coverage %, doc count, recent changes
"""

from typing import Optional
from datetime import date

from fastapi import APIRouter, Query, HTTPException, status, Depends
from pydantic import BaseModel, Field

from src.api.schemas import (
    CapabilitiesResponse,
    SearchLegalRequest,
    SearchLegalResponse,
    ResearchTaskRequest,
    ResearchTaskResponse,
    DocumentResponse,
    CitationGraphResponse,
    LegalUpdatesResponse,
    JurisdictionSummaryResponse,
)
from src.api.middleware.jwt_auth import get_current_tenant

router = APIRouter()


@router.get("/get_capabilities", response_model=CapabilitiesResponse)
async def get_capabilities(
    tenant: dict = Depends(get_current_tenant),
) -> CapabilitiesResponse:
    """Get system capabilities — full jurisdiction + doc type manifest.
    
    Returns all supported jurisdictions, document types, tools, and BAM tiers.
    """
    # TODO: Implement actual capabilities retrieval from database/config
    return CapabilitiesResponse(
        jurisdictions={
            "J_US_FED": {"name": "US Federal", "coverage_percent": 0.95, "doc_count": 150000},
            "J_CA_FED": {"name": "Canada Federal", "coverage_percent": 0.88, "doc_count": 45000},
            "J_EU": {"name": "European Union", "coverage_percent": 0.82, "doc_count": 78000},
        },
        doc_types=["STATUTE", "REGULATION", "CASE", "ADMINISTRATIVE_DECISION"],
        tools=[
            "get_capabilities",
            "search_legal",
            "research_task",
            "get_document",
            "get_citations",
            "check_updates",
            "jurisdiction_summary",
        ],
    )


@router.post("/search_legal", response_model=SearchLegalResponse)
async def search_legal(
    request: SearchLegalRequest,
    tenant: dict = Depends(get_current_tenant),
) -> SearchLegalResponse:
    """Search legal documents — hybrid pgvector + full-text search with BAM routing.
    
    Combines vector similarity search (HNSW index) with full-text search (GIN index)
    for optimal legal document retrieval.
    """
    # TODO: Implement hybrid search with pgvector and full-text
    # 1. Decompose query using BAM compound
    # 2. Vector search on legal_chunks.embedding (HNSW, cosine similarity)
    # 3. Full-text search on legal_chunks.chunk_text (GIN, tsvector)
    # 4. Re-rank and combine results
    # 5. Check query_cache for fingerprint match
    return SearchLegalResponse(
        results=[],
        total=0,
        latency_ms=0.0,
        cache_hit=False,
    )


@router.post("/research_task", response_model=ResearchTaskResponse)
async def research_task(
    request: ResearchTaskRequest,
    tenant: dict = Depends(get_current_tenant),
) -> ResearchTaskResponse:
    """Execute research task — complex research with question decomposition.
    
    Breaks down complex questions into sub-questions, runs parallel searches,
    and synthesizes a structured report with citation chain.
    """
    # TODO: Implement research pipeline
    # 1. Question decomposition via LLM
    # 2. Parallel search_legal calls per sub-question
    # 3. Synthesis of results into structured report
    # 4. Citation chain extraction
    # 5. Gap detection
    return ResearchTaskResponse(
        report="",
        citation_chain=[],
        confidence=0.0,
        gap_detected=False,
    )


@router.get("/get_document", response_model=DocumentResponse)
async def get_document(
    doc_id: Optional[str] = Query(None, description="Document UUID"),
    citation: Optional[str] = Query(None, description="Document citation"),
    tenant: dict = Depends(get_current_tenant),
) -> DocumentResponse:
    """Get document by ID or citation — retrieve full document with chunks and citations.
    
    Returns the complete document including all chunks, citation graph,
    and metadata. Requires either doc_id or citation parameter.
    """
    if not doc_id and not citation:
        raise HTTPException(
            status_code=status.HTTP_400_BAD_REQUEST,
            detail="Either doc_id or citation must be provided",
        )
    
    # TODO: Implement document retrieval from database
    return DocumentResponse(
        document={},
        chunks=[],
        citations=[],
    )


@router.get("/get_citations", response_model=CitationGraphResponse)
async def get_citations(
    doc_id: str = Query(..., description="Document UUID"),
    direction: str = Query("both", pattern=r"^(forward|backward|both)$"),
    depth: int = Query(3, ge=1, le=5),
    tenant: dict = Depends(get_current_tenant),
) -> CitationGraphResponse:
    """Get citation graph — retrieve citation graph up to specified depth.
    
    Traverses the legal_citations table to build a citation graph
    from the specified document, following forward (cited by),
    backward (cites), or both directions.
    """
    # TODO: Implement citation graph traversal
    # 1. BFS/DFS traversal of legal_citations table
    # 2. Build node and edge lists
    # 3. Extract authority chain
    # 4. Detect overruled cases
    return CitationGraphResponse(
        nodes=[],
        edges=[],
        authority_chain=[],
        overruled_cases=[],
    )


@router.get("/check_updates", response_model=LegalUpdatesResponse)
async def check_updates(
    jurisdiction: str = Query(..., description="Jurisdiction code"),
    since_date: date = Query(..., description="Date to check changes since"),
    tenant: dict = Depends(get_current_tenant),
) -> LegalUpdatesResponse:
    """Check for legal updates — all legislative changes since specified date.
    
    Queries the legal_updates table for changes detected by the
    monitor system since the given date for the specified jurisdiction.
    """
    # TODO: Implement update checking from legal_updates table
    return LegalUpdatesResponse(updates=[])


@router.get("/jurisdiction_summary", response_model=JurisdictionSummaryResponse)
async def jurisdiction_summary(
    jurisdiction_code: str = Query(..., description="Jurisdiction code"),
    tenant: dict = Depends(get_current_tenant),
) -> JurisdictionSummaryResponse:
    """Get jurisdiction summary — coverage %, doc count, recent changes.
    
    Aggregates statistics from legal_documents and legal_updates tables
    for the specified jurisdiction.
    """
    # TODO: Implement summary aggregation from database
    return JurisdictionSummaryResponse(
        jurisdiction_code=jurisdiction_code,
        jurisdiction_name="",
        coverage_percent=0.0,
        doc_count=0,
        recent_changes=[],
    )