# docs/openapi.yaml
openapi: 3.1.0
info:
  title: LexCore + LexRadar API
  description: |
    Complete API specification for LexCore legal intelligence database and LexRadar IP detection system.
    
    ## Authentication
    All endpoints require authentication via either:
    - JWT Bearer token (access + refresh tokens)
    - API key (tenant-specific)
    
    ## Rate Limiting
    Rate limits are enforced per-tenant based on tier:
    - Solo: 10K queries/month
    - Firm: Unlimited
    - Enterprise: Unlimited + self-host
    
    ## BAM Routing
    All requests include BAM compound signals for agent routing.
  version: 1.0.0
  contact:
    name: LexCore API Team
  license:
    name: Proprietary
servers:
  - url: https://api.lexcore.io/v1
    description: Production server
  - url: https://staging.lexcore.io/v1
    description: Staging server
  - url: http://localhost:8000/v1
    description: Local development
security:
  - BearerAuth: []
  - ApiKeyAuth: []
...
paths:
...
  /mcp/search_legal:
    post:
      tags:
        - mcp
      summary: Search legal documents
      description: Hybrid search combining pgvector + full-text search with BAM routing
      operationId: searchLegal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SearchLegalRequest'
      responses:
        '200':
          description: Search results returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchLegalResponse'
  /mcp/research_task:
    post:
      tags:
        - mcp
      summary: Execute research task
      description: Complex research with question decomposition and parallel search
      operationId: researchTask
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResearchTaskRequest'
      responses:
        '200':
          description: Research report generated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResearchTaskResponse'
  /mcp/get_document:
    get:
      tags:
        - mcp
      summary: Get document by ID or citation
      description: Retrieve full document with all chunks and citations
      operationId: getDocument
      parameters:
        - name: doc_id
          in: query
          schema:
            type: string
          description: Document UUID
        - name: citation
          in: query
          schema:
            type: string
          description: Document citation
      responses:
        '200':
          description: Document retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentResponse'
  /mcp/get_citations:
    get:
      tags:
        - mcp
      summary: Get citation graph
      description: Retrieve citation graph up to specified depth
      operationId: getCitations
      parameters:
        - name: doc_id
          in: query
          required: true
          schema:
            type: string
        - name: direction
          in: query
          schema:
            type: string
            enum: [forward, backward, both]
            default: both
        - name: depth
          in: query
          schema:
            type: integer
            default: 3
            maximum: 5
      responses:
        '200':
          description: Citation graph retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CitationGraphResponse'
  /mcp/check_updates:
    get:
      tags:
        - mcp
      summary: Check for legal updates
      description: Returns all legislative changes since specified date
      operationId: checkUpdates
      parameters:
        - name: jurisdiction
          in: query
          required: true
          schema:
            type: string
        - name: since_date
          in: query
          required: true
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Updates retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LegalUpdatesResponse'
  /mcp/jurisdiction_summary:
    get:
      tags:
        - mcp
      summary: Get jurisdiction summary
      description: Returns coverage %, doc count, recent changes
      operationId: jurisdictionSummary
      parameters:
        - name: jurisdiction_code
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Summary retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JurisdictionSummaryResponse'

  # ============================================================================
  # LEXCORE ENDPOINTS
  # ============================================================================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
...
    SearchLegalRequest:
      type: object
      required:
        - query
      properties:
        query:
          type: string
        filters:
          type: object
          properties:
            jurisdiction:
              type: string
            body_of_law:
              type: string
            doc_type:
              type: string
        top_k:
          type: integer
          default: 10
          maximum: 100
        config:
          type: object
          properties:
            weight_vector:
              type: number
              default: 0.7
            weight_fulltext:
              type: number
              default: 0.3
        bam_compound:
          type: string
          description: BAM routing compound
    SearchLegalResponse:
      type: object
      properties:
        results:
          type: array
          items:
            type: object
            properties:
              chunk_id:
                type: string
              document_id:
                type: string
              document_title:
                type: string
              chunk_text:
                type: string
              confidence:
                type: number
              freshness:
                type: string
              bam_signal:
                type: string
        total:
          type: integer
        latency_ms:
          type: number
        cache_hit:
          type: boolean
    ResearchTaskRequest:
      type: object
      required:
        - question
      properties:
        question:
          type: string
        jurisdictions:
          type: array
          items:
            type: string
        depth:
          type: integer
          default: 3
          maximum: 5
        bam_compound:
          type: string
    ResearchTaskResponse:
      type: object
      properties:
        report:
          type: string
        citation_chain:
          type: array
          items:
            type: object
        confidence:
          type: number
        gap_detected:
          type: boolean
        gap_description:
          type: string
    DocumentResponse:
      type: object
      properties:
        document:
          type: object
          properties:
            id:
              type: string
            title:
              type: string
            citation:
              type: string
            body_of_law:
              type: string
            jurisdiction:
              type: string
            full_text:
              type: string
            effective_date:
              type: string
            is_repealed:
              type: boolean
        chunks:
          type: array
          items:
            type: object
        citations:
          type: array
          items:
            type: object
    CitationGraphResponse:
      type: object
      properties:
        nodes:
          type: array
          items:
            type: object
        edges:
          type: array
          items:
            type: object
        authority_chain:
          type: array
          items:
            type: string
        overruled_cases:
          type: array
          items:
            type: string
    LegalUpdatesResponse:
      type: object
      properties:
        updates:
          type: array
          items:
            type: object
            properties:
              document_id:
                type: string
              update_type:
                type: string
              change_summary:
                type: string
              detected_at:
                type: string
    JurisdictionSummaryResponse:
      type: object
      properties:
        jurisdiction_code:
          type: string
        jurisdiction_name:
          type: string
        coverage_percent:
          type: number
        doc_count:
          type: integer
        recent_changes:
          type: array
          items:
            type: object

    # ============================================================================
    # LEXCORE SCHEMAS
    # ============================================================================
...